+++ askbot/conf/sidebar_question.py
@@ -16,9 +16,10 @@ settings.register(
     values.LongStringValue(
         SIDEBAR_QUESTION,
         'QUESTION_PAGE_TOP_BANNER',
-        description = _('Top banner'),
-        default = '',
-        help_text = _(
+        description=_('Top banner'),
+        default='',
+        localized=True,
+        help_text=_(
             'When using this option, please '
             'use the HTML validation service to make sure that '
             'your input is valid and works well in all browsers.'
@@ -39,9 +40,10 @@ settings.register(
     values.LongStringValue(
         SIDEBAR_QUESTION,
         'QUESTION_PAGE_ANSWER_BANNER',
-        description = _('Answers banner'),
-        default = '',
-        help_text = _(
+        description=_('Answers banner'),
+        default='',
+        localized=True,
+        help_text=_(
             'This banner will show under the first answer. '
             'When using this option, please '
             'use the HTML validation service to make sure that '
@@ -73,10 +75,11 @@ settings.register(
     values.LongStringValue(
         SIDEBAR_QUESTION,
         'SIDEBAR_QUESTION_HEADER',
-        description = _('Custom sidebar header'),
-        default = '',
-        help_text = _(
-                    'Use this area to enter content at the TOP of the sidebar'
+        description=_('Custom sidebar header'),
+        default='',
+        localized=True,
+        help_text=_(
+                    'Use this area to enter content at the TOP of the sidebar '
                     'in HTML format. When using this option '
                     '(as well as the sidebar footer), please '
                     'use the HTML validation service to make sure that '
@@ -138,9 +141,10 @@ settings.register(
     values.LongStringValue(
         SIDEBAR_QUESTION,
         'SIDEBAR_QUESTION_FOOTER',
-        description = _('Custom sidebar footer'),
-        default = '',
-        help_text = _(
+        description=_('Custom sidebar footer'),
+        default='',
+        localized=True,
+        help_text=_(
                     'Use this area to enter content at the BOTTOM of the sidebar'
                     'in HTML format.   When using this option '
                     '(as well as the sidebar header), please '