+++ askbot/models/__init__.py
@@ -787,6 +787,14 @@ def user_assert_can_upload_file(request_user):
     )
 
 
+def user_assert_can_merge_questions(self):
+    _assert_user_can(
+        user=self,
+        action_display=_('merge duplicate questions'),
+        admin_or_moderator_required=True
+    )
+
+
 def user_assert_can_post_text(self, text):
     """Raises exceptions.PermissionDenied, if user does not have
     privilege to post given text, depending on the contents
@@ -854,6 +862,10 @@ def user_assert_can_edit_comment(self, comment = None):
                 return
             else:
                 return
+
+    if not (self.is_blocked() or self.is_suspended()):
+        if self.reputation >= askbot_settings.MIN_REP_TO_EDIT_OTHERS_POSTS:
+            return
 
     error_message = _(
         'Sorry, but only post owners or moderators can edit comments'
@@ -1010,16 +1022,20 @@ def user_assert_can_delete_question(self, question = None):
             if self.is_administrator() or self.is_moderator():
                 return
             else:
+                if answer_count > 1:
+                    upvoted_answers_phrase = askbot_settings.WORDS_UPVOTED_ANSWERS
+                else:
+                    upvoted_answers_phrase = askbot_settings.WORDS_UPVOTED_ANSWER
+
                 msg = ungettext(
                     'Sorry, cannot %(delete_your_question)s since it '
-                    'has an %(upvoted_answer)s posted by someone else',
+                    'has an %(upvoted_answers)s posted by someone else',
                     'Sorry, cannot %(delete_your_question)s since it '
                     'has some %(upvoted_answers)s posted by other users',
                     answer_count
                 ) % {
                     'delete_your_question': askbot_settings.WORDS_DELETE_YOUR_QUESTION,
-                    'upvoted_answer': askbot_settings.WORDS_UPVOTED_ANSWER,
-                    'upvoted_answers': askbot_settings.WORDS_UPVOTED_ANSWERS
+                    'upvoted_answers': upvoted_answers_phrase
                 }
                 raise django_exceptions.PermissionDenied(msg)
 
@@ -1385,6 +1401,38 @@ def user_mark_tags(
 
     return cleaned_tagnames, cleaned_wildcards
 
+def user_merge_duplicate_questions(self, from_q, to_q):
+    """merges content from the ``from_thread`` to the ``to-thread``"""
+    #todo: maybe assertion will depend on which questions are merged
+    self.assert_can_merge_questions()
+    to_q.merge_post(from_q)
+    from_thread = from_q.thread
+    to_thread = to_q.thread
+    #set new thread value to all posts
+    posts = from_thread.posts.all()
+    posts.update(thread=to_thread)
+
+    if askbot_settings.LIMIT_ONE_ANSWER_PER_USER:
+        #merge answers if only one is allowed per user
+        answers = to_thread.all_answers()
+        answer_map = collections.defaultdict(list)
+        #compile all answers by user
+        for answer in answers:
+            author = answer.author
+            answer_map[author].append(answer)
+
+        for author in answer_map:
+            author_answers = answer_map[author]
+            if author_answers > 1:
+                first_answer = author_answers.pop(0)
+                for answer in author_answers:
+                    first_answer.merge_post(answer)
+
+    #from_thread.spaces.clear()
+    from_thread.delete()
+    to_thread.invalidate_cached_data()
+
+
 @auto_now_timestamp
 def user_retag_question(
                     self,
@@ -1540,7 +1588,8 @@ def user_delete_question(
     question.thread.save()
 
     for tag in list(question.thread.tags.all()):
-        if tag.used_count == 1:
+        if tag.used_count <= 1:
+            tag.used_count = 0
             tag.deleted = True
             tag.deleted_by = self
             tag.deleted_at = timestamp
@@ -1572,7 +1621,9 @@ def user_delete_all_content_authored_by_user(self, author, timestamp=None):
 
     #delete questions
     questions = Post.objects.get_questions().filter(author=author)
-    count += questions.update(deleted_at=timestamp, deleted_by=self, deleted=True)
+    count += questions.count()
+    for question in questions:
+        self.delete_question(question=question, timestamp=timestamp)
 
     threads = Thread.objects.filter(last_activity_by=author)
     for thread in threads:
@@ -1590,6 +1641,14 @@ def user_delete_all_content_authored_by_user(self, author, timestamp=None):
     comments = Post.objects.get_comments().filter(author=author)
     count += comments.count()
     comments.delete()
+
+    #delete all unused tags created by this user
+    #tags = author.created_tags.all()
+    #tag_ids = list()
+    #for tag in tags:
+    #    if tag.used_count == 0:
+    #        tag_ids.append(tag.id)
+    #Tag.objects.filter(id__in=tag_ids).delete()
 
     return count
 
@@ -2274,7 +2333,7 @@ def user_get_status_display(self):
     if self.is_approved():
         return _('Registered User')
     elif self.is_administrator():
-        return _('Adminstrator')
+        return _('Administrator')
     elif self.is_moderator():
         return _('Moderator')
     elif self.is_suspended():
@@ -3018,6 +3077,7 @@ User.add_to_class('follow_question', user_follow_question)
 User.add_to_class('unfollow_question', user_unfollow_question)
 User.add_to_class('is_following_question', user_is_following_question)
 User.add_to_class('mark_tags', user_mark_tags)
+User.add_to_class('merge_duplicate_questions', user_merge_duplicate_questions)
 User.add_to_class('update_response_counts', user_update_response_counts)
 User.add_to_class('can_create_tags', user_can_create_tags)
 User.add_to_class('can_have_strong_url', user_can_have_strong_url)
@@ -3076,6 +3136,7 @@ User.add_to_class('is_read_only', user_is_read_only)
 User.add_to_class('assert_can_vote_for_post', user_assert_can_vote_for_post)
 User.add_to_class('assert_can_revoke_old_vote', user_assert_can_revoke_old_vote)
 User.add_to_class('assert_can_upload_file', user_assert_can_upload_file)
+User.add_to_class('assert_can_merge_questions', user_assert_can_merge_questions)
 User.add_to_class('assert_can_post_question', user_assert_can_post_question)
 User.add_to_class('assert_can_post_answer', user_assert_can_post_answer)
 User.add_to_class('assert_can_post_comment', user_assert_can_post_comment)
@@ -3685,6 +3746,7 @@ def greet_new_user(user, **kwargs):
         template_name = 'email/welcome_lamson_off.html'
 
     data = {
+        'recipient_user': user,
         'site_name': askbot_settings.APP_SHORT_NAME,
         'site_url': site_url(reverse('questions')),
         'ask_address': 'ask@' + askbot_settings.REPLY_BY_EMAIL_HOSTNAME,